backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
langchain==0.1.0
langchain-community==0.0.10
sentence-transformers==2.2.2
gpt4all==2.0.2
faiss-cpu==1.7.4
PyPDF2==3.0.1
pdfplumber==0.10.3
python-pptx==0.6.23
python-docx==1.1.0
pydantic==2.5.0
python-multipart==0.0.6
numpy==1.24.3
